Data Miner v0.1.0 DataMiner.Apriori

Documentation for Apriori Algorithm Implementation.

Functions

apriori(frequencies, frequents, transactions, min_supp, transactions_length)

Implementation of eclat algorithm, this function will return any frequent itemset.

Examples

iex> DataMiner.Eclat.eclat([{[:a], MapSet.new([2])}, {[:b], MapSet.new([2])}], [], 0.1, 3)
[[{[:b, :a], #MapSet<[2]>}], [{[:a], #MapSet<[2]>}, {[:b], #MapSet<[2]>}]]

calculate_itemsets_frequency(itemsets, transactions)

This function will calculate frequency of any itemset by see itemset frequency inside transactions.

Examples

iex> DataMiner.Apriori.calculate_itemsets_frequency([["a", "b"], ["a", "c"]] |> Enum.map(&MapSet.new(&1)), [["a", "c", "d"], ["b", "c", "e"], ["a", "b", "c", "e"], ["b", "e"]] |> Enum.map(&MapSet.new(&1)))
[{["a", "b"], 1}, {["a", "c"], 2}]

export_frequents(frequents)

Export frequents will export all frequent itemsets to a file.

import_file(file_address)

import file.

import_frequencies()

import frequencies file.

import_transactions()

import transactions file.

main(min_supp)

Main function for run algorithm with minimum support.

This function will get minimum support as input.

This number is expressed as a percentage.

At the end of function result of Apriori algorithm will save to a file.

merge_itemsets(list, merged_itemsets)

This function will merge a list of itemsets to a list of sub itemsets. So input is a list of itemsets and output is a list of merged itemsets.

merger(list1, list2)

merger will merge two itemsets.

## Examples

iex> DataMiner.Apriori.merger([1, 2, 3], [4, 2, 3])
[4, 1, 2, 3]

remove_low_frequencies(transactions_length, frequencies, min_supp)

When itemsets merged succesfully, we should pipe them into remove_low_frequencies that will remove all of itemsets that size of their transactions are lower that minimum support.

This is for downward closers!

support(item_frequency, transactions_length)

support will calculate support of an itemset by its frequency