Data Miner v0.1.0 DataMiner.Eclat

Documentation for Eclat Algorithm Implementation.

Functions

eclat(itemsets, frequents, min_supp, transactions_length)

Implementation of eclat algorithm, this function will return any frequent itemset.

Examples

iex> DataMiner.Eclat.eclat([{[:a], MapSet.new([2])}, {[:b], MapSet.new([2])}], [], 0.1, 3)
[[{[:b, :a], #MapSet<[2]>}], [{[:a], #MapSet<[2]>}, {[:b], #MapSet<[2]>}]]

export_frequents(frequents)

Export frequents will export all frequent itemsets to a file.

import_file(file_address)

Import file.

import_transactions()

import transactions file.

main(min_supp)

Main function for run algorithm with minimum support.

This function will get minimum support as input.

This number is expressed as a percentage.

At the end of function result of Eclat algorithm will save to a file.

merge_itemsets(itemsets)

This function will merge a list of itemsets to a list of sub itemsets. So input is a list of itemsets and output is a list of merged itemsets.

note: Commented codes are a parallel code for merging.

Examples

iex> DataMiner.Eclat.merge_itemsets([{[2, 1], MapSet.new([2])}, {[3, 1], MapSet.new([2])}])
[{[3, 2, 1], #MapSet<[2]>}]

merger(arg1, arg2, group_tail)

This function will merge an itemset with another itemset.

What is merge itemsets and make sub itemset? if we have a = [1, 2, 4] and b = [1, 2, 5] then merge of them will be: result = [1, 2, 4, 5]

a and b can merge because of (0 .. k-1)th items in their lists are similar.

In this module for avoiding of list overhead, we merge lists by (1 .. k)th items. if a = [2, 1] and b = [3, 1] then merge of them with this algorithm will be: result = [3, 2, 1]

Examples

iex> DataMiner.Eclat.merge_itemsets([{[2, 1], MapSet.new([2])}, {[3, 1], MapSet.new([2])}])
[{[3, 2, 1], #MapSet<[2]>}]

remove_low_frequencies(itemsets, min_supp, transactions_length)

When itemsets merged succesfully, we should pipe them into remove_low_frequencies that will remove all of itemsets that size of their transactions are lower that minimum support.

This is for downward closers!

Examples

iex> DataMiner.Eclat.remove_low_frequencies([{[:a], MapSet.new([1, 2, 3])}, {[:b], MapSet.new([1])}], 50, 3)
[{[:a], #MapSet<[1, 2, 3]>}]

support(item_frequency, transactions_length)

support will calculate support of an itemset by its frequency

transactios_to_eclat_form(transactions)

This function will get main transactions and return eclat form of that.

Eclat form is a map that show transactions of an item that it is inside them!

Examples

iex> DataMiner.Eclat.transactios_to_eclat_form([["1", "2"], ["2", "4"], ["1", "5"], ["1", "6", "7", "3", "1", "2", "9"]])
{[:"1"] => #MapSet<[0, 2, 3]>, [:"2"] => #MapSet<[0, 1, 3]>, [:"3"] => #MapSet<[3]>, [:"4"] => #MapSet<[1]>, [:"5"] => #MapSet<[2]>, [:"6"] => #MapSet<[3]>, [:"7"] => #MapSet<[3]>, [:"9"] => #MapSet<[3]>}